Introduce `ErrorProneTestHelperSourceFormat` check
This new checker inspects inline code passed to `CompilationTestHelper`
and `BugCheckerRefactoringTestHelper`. It requires that this code is
properly formatted and that its imports are organized. Only code that
represents the expected output of a refactoring operation is allowed to
have unused imports, as most `BugChecker`s do not (and are not able to)
remove imports that become obsolete as a result of applying their
suggested fix(es).