Fix Javadoc JAR generation
For performance reasons the Javadoc JARs are not generated when
executing `mvn install`. They are generally only compiled when running
`mvn release:perform`, as part of the `release` Maven profile.

The downside of this setup is that Javadoc generation issues may not be
caught until release time. This change resolves such an issue. To
prevent future regressions Javadoc generation is now also performed by
the GitHub Actions build workflow.

While there, drop the deprecated `useReleaseProfile` configuration
setting of the javadoc-maven-plugin, as its value matches the default.