Prefer `Flux#take(n, true)` over `Flux#take(n)` to limit generation
In Reactor versions prior to 3.5.0, `Flux#take(n)` makes an unbounded
request upstream, and is equivalent to `Flux#take(n, false)`.
In 3.5.0, the behavior of `Flux#take(n)` will change to that of
`Flux#take(n, true)`.

The intent with this Refaster rule is to get the new behavior
before upgrading to Reactor 3.5.0.